refactor: Operator 연산자 클래스를 추가하여 유효성 로직을 Calculator로부터 덜어 냄
